Modifed sendmsg_withlength and recvmsg_withlength to use void pointers
these functions were designed to send raw data. Character arrays are not
meant to be raw data, while a void* pointer clearly shows the intention
that the data could be whatever you want. This should have no effect on
the result, because the pointers can always be type cast, but
it helps understand what we are doing with the information