Unstable API changes: modifying send+recv
The send and recv functions as designed do not need to use the convoluted loops
that are currently in place, because send and recv will block until
the complete send is finished, as long as the flag DONT_WAIT is not enabled.

We do not want this flag enabled, and we don't want to keep attempting a failed send,
so it would be best to check errors and simply return if the send does not complete
normally.