Modified opt.c and Makefile
Added libopt to the makefile, also fixed a problem with the default
paramater.

C++ has the default option feature, but apparently C does not :(

So instead, I changed opt.h to use a global const string,
and have filename point at that string if it is null.

Note that in using the function you have to put null if you
want to use the default filename.